+++ CHANGES.md
@@ -5,10 +5,12 @@ Here you can see the full list of changes between each PyScript release.
 
 Version 0.2.5
 -----------
+- changed: `requests.interceptors` must return explicitly false to exit handling of response
+- changed: `requests.interceptors` now takes an object which can be for requests as well
 - changed: renamed `list.invoke` with `list.map`
-- fixed: `request.upload` can now specify headers
+- fixed: `requests.upload` can now specify headers
 - removed: `str.sprintf`, which wasn't working in previous version
-- removed: `request.beforeRequest`, replaced by whenGET, whenPOST, etc.
+- removed: `requests.beforeRequest`, replaced by `requests.interceptors`
 
 Version 0.2.4
 -----------